move default seccomp profile into package
